Basic Fielding Code
This is pretty good so far.
I need to make this happen like 10 or so times to make it like actual players on a baseball field.
That however is a problem for tomorrow.
I also need to add in scrolling so the screen follows the ball at all cost. Also I need to figure out how to like make the ball do the thing so that it follows like a decent trajectory when you like hit it????? Idk how I'm gonna do that but that's a can to kick down the road for later.
There's also the matter of making the players actually follow the boundaries of the baseball field but first I need a baseball field to follow. Also AI needs to happen which will be an absolute nightmare.
This project is like 30 minutes old and is already nigh impossible but I'm bored enough to make it happen.

I am tired goodnight